Added mouseup and mousedown functions on body
Previously, when creating walls, it seemed to be kind of buggy. Also, when the user drags out of the page, on bringing the cursor back, it would continue to create walls. Now, mouseup function has been added to body to reduce buggy feeling and also, mouseleave function has been added to stop creating walls after leaving the page and returning back. NOTE: This was only done for normal elements and NOT special elements.